bind문제 해결했습니다
